Extract membership admin card to own component
